==> Essential_DAmP_screen/src/00_createYeastGenesets.Rmd <==
---
title: "S. cerevisiae genesets"
author: "Ashwini Kumar Sharma, PhD"
date: "`r Sys.Date()`"
output:
  rmdformats::readthedown:
    highlight: kate
    code_folding: hide
---

```{r knitr_init, echo=FALSE, cache=FALSE}
library(knitr)
library(rmdformats)

## Global options
options(max.print=100000)
opts_chunk$set(echo=TRUE,
	             cache=FALSE,
               prompt=FALSE,
               tidy=TRUE,
               comment=NA,
               message=FALSE,
               warning=FALSE)
opts_knit$set(width=80)
```

## Initialization

We start the analysis by initializing the packages required in this section. We also define the root directory, within which all the input/output operations for this project will be performed. At the [end of this document](#SI), a detailed software version information is provided for easier reproducibility of the analysis.

```{r}
library(data.table)
library(gridExtra)
library(KEGGREST)
library(ggplot2)
library(reshape2)
path = "/Users/ashwin/Documents/Projects/YeastScreen/Essential_DAmP_screen/"
```

***

## Slim gene ontology

Creating yeast (_Saccharomyces cerevisiae_) specific slim gene ontology based genesets. We create separate sets for `molecular function` and `biological process`. The data is obtained from [Saccharomyces Genome Database (SGD)](https://www.yeastgenome.org/). Keeping only those genesets having >= 5 and <= 500 genes.

```{r}

dat  = read.delim(
  "https://downloads.yeastgenome.org/curation/literature/go_slim_mapping.tab",
  header = F,
  stringsAsFactors = F
)
colnames(dat) = c("ORF", "Gene", "SGDID", "GO_Aspect", "GOslim", "GOID", "Type")

bp = dat[dat$GO_Aspect == "P", ]
mf = dat[dat$GO_Aspect == "F", ]

createGoObj = function(godat)
{
  godat = split(godat$Gene, godat$GOslim)
  godat = lapply(godat, function(x) {
    x = x[x != ""]
    x = unique(x)
    return(x)
  })
  
  return(godat)
}

bp = createGoObj(godat = bp)
cnt = sapply(bp, length)
bp = bp[ cnt >=5 & cnt <= 500]

mf = createGoObj(godat = mf)
cnt = sapply(mf, length)
mf = mf[cnt >=5 & cnt <= 500]

```

***

## KEGG pathways

Creating yeast (_Saccharomyces cerevisiae_) specific KEGG pathway based genesets. The data is obtained from [Saccharomyces Genome Database (SGD)](https://www.yeastgenome.org/). Keeping only those gene sets having >= 5 and <= 500 genes.

```{r}

# metpath = read.delim(url("https://downloads.yeastgenome.org/curation/literature/biochemical_pathways.tab"), header = F, stringsAsFactors = F)
# colnames(metpath) = c("Pathway", "Enzyme", "ECnumber", "Gene", "Reference")
# metpath = metpath[- which(metpath$Gene == ""),]
# metpath = split(metpath$Gene, metpath$Pathway)
# metpath = metpath[sapply(metpath, length) >= 3]
# yeastpath = metpath
# rm(metpath)

## Or KEEG pathways can be extracted from KEEG website directly

yeastpath = keggLink("pathway", "sce")
yeastpath = unique(yeastpath)

ind = seq(1, length(yeastpath), 10)
ind = data.frame(start = ind, end = c(c(ind[2:length(ind)] - 1), length(yeastpath)))

pathres = vector("list", length(yeastpath))

for(i in 1:nrow(ind))
{
  tmp = keggGet(yeastpath[ ind[i,1] : ind[i,2] ])
  pathres[ind[i,1]:ind[i,2]] = lapply(tmp, function(x){
    x = grep(";", x$GENE, value=T)
    x = unique(sapply(strsplit(x , ";"), function(y)y[1]))
    return(x)
  })
  names(pathres)[ind[i,1]:ind[i,2]] = gsub(" - Saccharomyces cerevisiae (budding yeast)", "", sapply(tmp, function(x){x=x$NAME}), fixed=T)
  rm(tmp)
}
yeastpath = pathres
rm(i, ind, pathres)

# Keeping only the pathways with at least 5 genes
cnt = sapply(yeastpath, length)
yeastpath = yeastpath[cnt >= 5 & cnt <= 500]

```

***

## Transcription fator targets

Creating yeast (_Saccharomyces cerevisiae_) specific transcription factor targets. The data is originally obtained from [TF2DNA database](http://www.fiserlab.org/tf2dna_db/). However, we use the data provided from [YeastEnrichr](https://amp.pharm.mssm.edu/YeastEnrichr/) which also based on TF2DNA but provides it in an easy-to-use format. Keeping only those gene sets having >= 5 and <= 1000 genes.

```{r}
if(!file.exists(paste0(path,"publicData/TF2DNA_2018.txt"))){
  download.file(url = "https://amp.pharm.mssm.edu/YeastEnrichr/geneSetLibrary?mode=text&libraryName=TF2DNA_2018", destfile = paste0(path,"publicData/TF2DNA_2018.txt"))
}

tfs = readLines(paste0(path,"publicData/TF2DNA_2018.txt"))
tfs = lapply(as.list(tfs), function(x) {
  x  = unique(unlist(strsplit(x, "\t", fixed = T)))
  x = x[!x == ""]
})
names(tfs) = sapply(tfs, function(x)x[1])
tfs = lapply(tfs, function(x)x[-1])

tfid = data.frame(study = names(tfs), tf = sapply(strsplit(names(tfs), "_"), function(x)x[1]), stringsAsFactors = F)
tfid = split(tfid$study, tfid$tf)

tfs = lapply(tfid, function(x){
  x = Reduce(intersect, tfs[x])
})

cnt = sapply(tfs, length)
tfs = tfs[cnt >=5 & cnt <=1000]

```

***

## Yeast phenotypes

Creating yeast (_Saccharomyces cerevisiae_) specific gene phenotype information. The data is obtained from [Saccharomyces Genome Database (SGD)](https://www.yeastgenome.org/). Keeping only those gene sets having >= 5 and<= 1000 genes.

```{r}
pheno  = read.delim(
  "https://downloads.yeastgenome.org/curation/literature/phenotype_data.tab",
  header = F,
  stringsAsFactors = F
)

pheno = pheno[,c(3,6,7,10)]
colnames(pheno) = c("Gene", "Experiment", "MutantType", "Phenotype")

pheno = lapply(split(pheno$Gene, pheno$Phenotype), function(x){x = x[x != ""]; x = unique(x)})
cnt = sapply(pheno, length)
pheno = pheno[cnt >=5 & cnt <=1000]

```

***

## Integrated yeast genesets

Here we combine the KEGG and GO slim genesets into one meta yeast geneset object

```{r}
genesets = list(KEGG = yeastpath, GOBP = bp, GOMF = mf, TFtargets = tfs, Phenotype = pheno)
saveRDS(genesets, file = paste0(path,"data/workspaces/YeastGeneSets.RDS"))
```

## Geneset description

The number of pathways from each of the databases we used. `ALL` is simply the combination of the first 3.

```{r, fig.width=2.5, fig.height=4.5, fig.align="center"}
tmp = lapply(genesets, length)
tmp = melt(tmp)

p = ggplot(tmp, aes(reorder(L1, value), value)) + theme_bw(base_size = 8) + 
    labs(x = "Databases", y="Number of terms or genesets or pathways") +
    geom_bar(stat="identity") + geom_text(aes(label=value), vjust=-0.3, size=3.5) +
    theme(axis.text.x = element_text(angle = 45, hjust = 0.5, vjust = 0.5))

ggsave(filename = paste0(path, "analysis/supplementary/Scerevisae_genesets.pdf"), 
       plot = p, width = 4, height = 4)

p

```

The size of each of the gene sets, i.e. the number of genes per gene sets. Not showing phenotypes as the list is too big to display.

```{r, fig.height = 25, fig.width = 12, fig.align="center"}

tmp = lapply(genesets$GOMF, length)
tmp = melt(tmp)

p1 = ggplot(tmp, aes(reorder(L1, value), value)) + theme_bw(base_size = 9) + 
labs(x = "", y="Number of genes") +
geom_bar(stat="identity") + coord_flip() +
theme(axis.text.x = element_text(angle = 90)) 

tmp = lapply(genesets$GOBP, length)
tmp = melt(tmp)

p2 = ggplot(tmp, aes(reorder(L1, value), value)) + theme_bw(base_size = 9) + 
labs(x = "", y="Number of genes") +
geom_bar(stat="identity") + coord_flip() +
theme(axis.text.x = element_text(angle = 90)) 

tmp = lapply(genesets$KEGG, length)
tmp = melt(tmp)

p3 = ggplot(tmp, aes(reorder(L1, value), value)) + theme_bw(base_size = 9) + 
labs(x = "", y="Number of genes") +
geom_bar(stat="identity") + coord_flip() +
theme(axis.text.x = element_text(angle = 90)) 

tmp = lapply(genesets$TFtargets, length)
tmp = melt(tmp)

p4 = ggplot(tmp, aes(reorder(L1, value), value)) + theme_bw(base_size = 8) + 
labs(x = "", y="Number of genes") +
geom_bar(stat="identity") + coord_flip() +
theme(axis.text.x = element_text(angle = 90)) 

p = grid.arrange(p1, p2, p3, p4, ncol = 2, nrow = 2)

ggsave(filename = paste0(path, "analysis/supplementary/Scerevisae_genesets_detailed.pdf"), 
       plot = p, width = 10, height = 20)

```

***

## Session Information {#SI}

```{r}
sessionInfo()
```
